# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Haïtien Nud Média API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  message?: string;
}

export interface Video {
  id: string;
  title: string;
  description: string;
  thumbnailUrl: string;
  videoUrl: string;
  category: string;
  durationSec: number;
  views: number;
  createdAt: string;
  isVip: boolean;
  published?: boolean;
}

export interface VideoInput {
  /**
   * @minLength 2
   * @maxLength 200
   */
  title: string;
  /**
   * @minLength 2
   * @maxLength 5000
   */
  description: string;
  thumbnailUrl: string;
  videoUrl: string;
  category: string;
  /** @minimum 0 */
  durationSec: number;
  isVip?: boolean;
  published?: boolean;
}

export interface Comment {
  id: string;
  videoId: string;
  userId?: string | null;
  body: string;
  anonymous: boolean;
  displayName: string;
  createdAt: string;
}

export type MePlan = (typeof MePlan)[keyof typeof MePlan];

export const MePlan = {
  free: "free",
  vip: "vip",
} as const;

export interface Me {
  id: string;
  email: string;
  displayName?: string | null;
  plan: MePlan;
  downloadsRemaining: number;
  ageConfirmed: boolean;
  isAdmin: boolean;
  subscriptionEndsAt?: string | null;
}

export interface Plan {
  id: string;
  name: string;
  description?: string;
  durationDays: number;
  priceCents: number;
  currency: string;
  features: string[];
}

export type SubscriptionPlan =
  (typeof SubscriptionPlan)[keyof typeof SubscriptionPlan];

export const SubscriptionPlan = {
  free: "free",
  vip: "vip",
} as const;

export interface Subscription {
  plan: SubscriptionPlan;
  status: string;
  startedAt?: string | null;
  endsAt?: string | null;
  planName?: string | null;
}

export type SupportTicketStatus =
  (typeof SupportTicketStatus)[keyof typeof SupportTicketStatus];

export const SupportTicketStatus = {
  open: "open",
  answered: "answered",
  closed: "closed",
} as const;

export interface SupportTicket {
  id: string;
  userId?: string | null;
  subject: string;
  message: string;
  reply?: string | null;
  status: SupportTicketStatus;
  createdAt: string;
}

export interface AdsConfig {
  provider: string;
  adToken?: string | null;
  preroll: boolean;
  postroll: boolean;
}

export interface AdminStats {
  totalUsers: number;
  activeVip: number;
  totalVideos: number;
  totalViews: number;
  totalDownloads: number;
  openTickets: number;
}

export type AdminUserPlan = (typeof AdminUserPlan)[keyof typeof AdminUserPlan];

export const AdminUserPlan = {
  free: "free",
  vip: "vip",
} as const;

export interface AdminUser {
  id: string;
  email: string;
  displayName?: string | null;
  plan: AdminUserPlan;
  blocked: boolean;
  isAdmin: boolean;
  ageConfirmed?: boolean;
  createdAt: string;
  subscriptionEndsAt?: string | null;
}

export type ListVideosParams = {
  q?: string;
  category?: string;
};

export type RequestDownload200 = {
  url: string;
  remaining: number;
};

export type CreateCommentBody = {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  body: string;
  anonymous?: boolean;
};

export type CreateCheckoutBody = {
  planId: string;
};

export type CreateCheckout200 = {
  url: string;
};

export type CreateTicketBody = {
  /**
   * @minLength 2
   * @maxLength 200
   */
  subject: string;
  /**
   * @minLength 2
   * @maxLength 5000
   */
  message: string;
};

export type GetAdsConfigParams = {
  videoId?: string;
};

export type AdminBlockUserBody = {
  blocked: boolean;
};

export type AdminReplyTicketBody = {
  /**
   * @minLength 1
   * @maxLength 5000
   */
  reply: string;
};