# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Haïtien Nud Média API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/checkout:
    post:
      operationId: createCheckout
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId]
              properties:
                planId: { type: string }
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url: { type: string }
...
  /admin/videos:
    get:
      operationId: adminListVideos
      tags: [admin]
      responses:
        "200":
          description: Videos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Video" }
    post:
      operationId: adminCreateVideo
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Video" }
  /admin/videos/{id}:
    patch:
      operationId: adminUpdateVideo
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Video" }
    delete:
      operationId: adminDeleteVideo
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
...
  /admin/users/{id}/block:
    post:
      operationId: adminBlockUser
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blocked]
              properties:
                blocked: { type: boolean }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
...
  /admin/tickets/{id}/reply:
    post:
      operationId: adminReplyTicket
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reply]
              properties:
                reply: { type: string, minLength: 1, maxLength: 5000 }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupportTicket" }
components:
  schemas:
...
    Video:
      type: object
      required: [id, title, description, thumbnailUrl, videoUrl, category, durationSec, views, createdAt, isVip]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        thumbnailUrl: { type: string }
        videoUrl: { type: string }
        category: { type: string }
        durationSec: { type: integer }
        views: { type: integer }
        createdAt: { type: string, format: date-time }
        isVip: { type: boolean }
        published: { type: boolean }
    VideoInput:
      type: object
      required: [title, description, thumbnailUrl, videoUrl, category, durationSec]
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string, minLength: 2, maxLength: 5000 }
        thumbnailUrl: { type: string }
        videoUrl: { type: string }
        category: { type: string }
        durationSec: { type: integer, minimum: 0 }
        isVip: { type: boolean }
        published: { type: boolean }
...
    SupportTicket:
      type: object
      required: [id, subject, message, status, createdAt]
      properties:
        id: { type: string }
        userId: { type: string, nullable: true }
        subject: { type: string }
        message: { type: string }
        reply: { type: string, nullable: true }
        status: { type: string, enum: [open, answered, closed] }
        createdAt: { type: string, format: date-time }
...
    AdminUser:
      type: object
      required: [id, email, plan, blocked, createdAt, isAdmin]
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string, nullable: true }
        plan: { type: string, enum: [free, vip] }
        blocked: { type: boolean }
        isAdmin: { type: boolean }
        ageConfirmed: { type: boolean }
        createdAt: { type: string, format: date-time }
        subscriptionEndsAt: { type: string, format: date-time, nullable: true }