# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Haïtien Nud Média API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserPlan } from "./adminUserPlan";

export interface AdminUser {
  id: string;
  email: string;
  displayName?: string | null;
  plan: AdminUserPlan;
  blocked: boolean;
  isAdmin: boolean;
  ageConfirmed?: boolean;
  createdAt: Date;
  subscriptionEndsAt?: Date | null;
}