# lib/api-zod/src/generated/types/supportTicket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Haïtien Nud Média API
 * OpenAPI spec version: 0.1.0
 */
import type { SupportTicketStatus } from "./supportTicketStatus";

export interface SupportTicket {
  id: string;
  userId?: string | null;
  subject: string;
  message: string;
  reply?: string | null;
  status: SupportTicketStatus;
  createdAt: Date;
}