# lib/api-zod/src/generated/types/videoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Haïtien Nud Média API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoInput {
  /**
   * @minLength 2
   * @maxLength 200
   */
  title: string;
  /**
   * @minLength 2
   * @maxLength 5000
   */
  description: string;
  thumbnailUrl: string;
  videoUrl: string;
  category: string;
  /** @minimum 0 */
  durationSec: number;
  isVip?: boolean;
  published?: boolean;
}