# lib/db/src/schema/videos.ts
import { pgTable, text, timestamp, boolean, integer, uuid } from "drizzle-orm/pg-core";

export const videosTable = pgTable("videos", {
  id: uuid("id").primaryKey().defaultRandom(),
  title: text("title").notNull(),
  description: text("description").notNull().default(""),
  thumbnailUrl: text("thumbnail_url").notNull(),
  videoUrl: text("video_url").notNull(),
  category: text("category").notNull().default("Général"),
  durationSec: integer("duration_sec").notNull().default(0),
  views: integer("views").notNull().default(0),
  isVip: boolean("is_vip").notNull().default(false),
  published: boolean("published").notNull().default(true),
  createdAt: timestamp("created_at").notNull().defaultNow(),
});

export type Video = typeof videosTable.$inferSelect;
export type InsertVideo = typeof videosTable.$inferInsert;